Overhaul of the new play modal to clean it up and modularize
