+++ src/pages/film-room/[video]/index.tsx
@@ -3,9 +3,9 @@ import { useSearchParams } from "next/navigation";
 import { useRouter } from "next/router";
 import { useEffect, useRef, useState } from "react";
 import Youtube, { type YouTubeEvent, type YouTubePlayer } from "react-youtube";
+import NewPlayModal from "~/components/add-play";
 import PageTitle from "~/components/page-title";
 import PlayIndex from "~/components/play-index";
-import PlayModal from "~/components/play-modal";
 import { useAuthContext } from "~/contexts/auth";
 import { useMobileContext } from "~/contexts/mobile";
 import { useIsDarkContext } from "~/pages/_app";
@@ -26,7 +26,7 @@ const FilmRoom = () => {
   const playerRef = useRef<HTMLDivElement | null>(null);
   const [activePlay, setActivePlay] = useState<PlayPreviewType | null>(null);
 
-  const [isPlayModalOpen, setIsPlayModalOpen] = useState<boolean>(false);
+  const [isNewPlayOpen, setIsNewPlayOpen] = useState<boolean>(false);
 
   const fetchVideo = async () => {
     const { data } = await supabase
@@ -77,7 +77,7 @@ const FilmRoom = () => {
 
   return (
     video && (
-      <div className="m-4 flex w-full flex-col items-center justify-center gap-3">
+      <div className="flex w-full flex-col items-center justify-center gap-2 p-4">
         <div className="flex flex-col items-center justify-center gap-1 text-center">
           <div
             className={`${
@@ -93,16 +93,12 @@ const FilmRoom = () => {
           </div>
           <PageTitle title={video.title} size="large" />
         </div>
-        <Divider
-          flexItem
-          variant="middle"
-          sx={{ marginLeft: "8px", marginRight: "8px", marginTop: "4px" }}
-        ></Divider>
-        <PlayModal
+        <Divider flexItem variant="middle" sx={{ marginTop: "8px" }} />
+        <NewPlayModal
           player={player}
           video={video}
-          isPlayModalOpen={isPlayModalOpen}
-          setIsPlayModalOpen={setIsPlayModalOpen}
+          isNewPlayOpen={isNewPlayOpen}
+          setIsNewPlayOpen={setIsNewPlayOpen}
         />
         {video.link && (
           <div ref={playerRef}>
